Update the open source ikec program to be compatible with the size_t changes made to libike.
git-svn-id: svn://svn.shrew.net/ike/head@168 1bb156eb-19af-db11-b919-0014c23ffdf2

+++ source/ikec/ikec.cpp
@@ -1084,7 +1084,7 @@ void _IKEC::run()
 			case IKEI_MSGID_STATUS:
 			{
 				char	txtmsg[ IKEI_MAX_BDATA + 1 ] = { 0 };
-				long	txtlen = IKEI_MAX_BDATA;
+				size_t	txtlen = IKEI_MAX_BDATA;
 
 				result = ikei.recv_msg_status( &status, txtmsg, txtlen );
 				if( result != IKEI_OK )
